requirements.txt
absl-py==0.6.1
astor==0.7.1
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
bleach==1.5.0
boto==2.49.0
boto3==1.9.169
botocore==1.12.169
bz2file==0.98
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
docutils==0.14
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
gast==0.2.0
gensim==3.7.3
grpcio==1.16.1
html5lib==0.9999999
idna==2.8
jmespath==0.9.4
kiwisolver==1.1.0
Markdown==3.0.1
mkl-fft==1.0.12
mkl-random==1.0.2
mock==2.0.0
nltk==3.4.3
numpy==1.16.4
pbr==5.1.1
protobuf==3.6.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
s3transfer==0.2.1
scipy==1.2.2
singledispatch==3.4.0.3
six==1.12.0
smart-open==1.8.4
subprocess32==3.5.3
tensorboard==1.8.0
tensorflow-gpu==1.8.0rc1
termcolor==1.1.0
tornado==5.1.1
urllib3==1.25.3
Werkzeug==0.14.1
